refactor: move gpx conversions to route_gpx.rs
