+++ api/domain/src/model/route/coordinate.rs
@@ -4,7 +4,6 @@ use std::iter::FromIterator;
 use geo::algorithm::haversine_distance::HaversineDistance;
 use getset::Getters;
 use itertools::Itertools;
-use num_traits::FromPrimitive;
 use polyline::{decode_polyline, encode_coordinates};
 use serde::{Deserialize, Serialize};
 
@@ -16,12 +15,12 @@ use crate::model::types::{Distance, Elevation, Latitude, Longitude, Polyline};
 #[derive(Clone, Debug, PartialEq, Getters, Deserialize, Serialize)]
 #[get = "pub"]
 pub struct Coordinate {
-    latitude: Latitude,
-    longitude: Longitude,
+    pub(super) latitude: Latitude,
+    pub(super) longitude: Longitude,
     #[serde(skip_serializing_if = "Option::is_none")]
-    elevation: Option<Elevation>,
+    pub(super) elevation: Option<Elevation>,
     #[serde(skip_serializing_if = "Option::is_none")]
-    distance_from_start: Option<Distance>,
+    pub(super) distance_from_start: Option<Distance>,
 }
 
 impl Coordinate {
@@ -69,21 +68,6 @@ impl TryFrom<geo::Coordinate<f64>> for Coordinate {
             elevation: None,
             distance_from_start: None,
         })
-    }
-}
-
-impl From<Coordinate> for gpx::Waypoint {
-    fn from(coord: Coordinate) -> Self {
-        let elevation = coord
-            .elevation
-            .map(|elev| elev.value())
-            .map(f64::from_i32)
-            .flatten();
-
-        let mut waypoint = Self::new(<(f64, f64)>::from(coord).into());
-        waypoint.elevation = elevation;
-
-        waypoint
     }
 }
