+++ api/domain/src/model/route/route_gpx.rs
@@ -2,13 +2,75 @@ use std::convert::TryFrom;
 use std::io::{Cursor, Read};
 use std::str::from_utf8;
 
+use itertools::Itertools;
+use num_traits::FromPrimitive;
 use quick_xml::events::Event;
 use quick_xml::{Reader, Writer};
 
 use route_bucket_utils::{ApplicationError, ApplicationResult};
 
-use crate::model::route::Route;
+use crate::model::route::{
+    coordinate::Coordinate, route_info::RouteInfo, segment_list::SegmentList, Route,
+};
 
+#[cfg(test)]
+use derivative::Derivative;
+
+impl From<Coordinate> for gpx::Waypoint {
+    fn from(coord: Coordinate) -> Self {
+        let elevation = coord
+            .elevation
+            .map(|elev| elev.value())
+            .map(f64::from_i32)
+            .flatten();
+
+        let mut waypoint = Self::new(<(f64, f64)>::from(coord).into());
+        waypoint.elevation = elevation;
+
+        waypoint
+    }
+}
+
+impl From<RouteInfo> for gpx::Metadata {
+    fn from(route_info: RouteInfo) -> Self {
+        Self {
+            name: Some(route_info.name),
+            description: None,
+            // TODO: ここにRouteBucketのリンクを入れられると良さそう
+            author: None,
+            links: vec![],
+            time: None,
+            keywords: None,
+            bounds: None,
+        }
+    }
+}
+
+impl From<SegmentList> for gpx::Track {
+    fn from(seg_list: SegmentList) -> Self {
+        let mut trk = Self::new();
+        trk.segments.push(gpx::TrackSegment::new());
+        trk.segments[0].points = seg_list
+            .segments
+            .into_iter()
+            .map(|seg| seg.into_iter())
+            .flatten()
+            .map(gpx::Waypoint::from)
+            .collect_vec();
+        trk
+    }
+}
+
+impl From<Route> for gpx::Gpx {
+    fn from(route: Route) -> Self {
+        gpx::Gpx {
+            version: gpx::GpxVersion::Gpx11,
+            metadata: Some(route.info.into()),
+            tracks: vec![route.seg_list.into()],
+            ..Default::default()
+        }
+    }
+}
 pub struct RouteGpx {
     name: String,
     data: Vec<u8>,
